Write output data also to a "notes" file
Implement the required functionality to write certain output
data generated by Chef class to a "notes" file (plain txt file),
besides printing that data to standard ouput (usually, a terminal).

When executing "tell_me_about" method (and probably others),
the result may be too big to handle or even read it properly in
a terminal, so the result needs to be written in a file too, so
the user can work with it in a more friendly way.