Implement matching by field-pattern method
Implement a method to search for recipes where the specified
"field" contains the specified "pattern", and use such method
in the "tell_me_about" method.

This method can be used for simple pattern searchs in almost
all fields of a recipe (title, url, ingredients), so it
improves the modularity and reusability of the code.