Fix "find_matching_ingredients" method
Turn the recipe_ingredients list in a single string, so the "any"
and "all" functions will work as intended. The goal is to find
if any (or all) strings in "matching_ingredients" are substrings of
the "recipe_ingredients". If this is done with lists, several
matches wont appear as such.

I.e: if matching_ingredients is ["eggs", "bacon"]
and recipe_ingredients is ["small eggs, "smoky bacon"], the following
call to "any" will return FALSE, because matching between lists only
works when the element X of a list is EXACTLY equal
to element Y in the other list:

    any(x in recipe_ingredients for x in matching_ingredients)

This return false because element "eggs" is not found
in recipe_ingredients list.

Thats the root cause: comparison between elements
instead of checking if substring in string.