Add example recipes datasets
Add an example of a recipe.csv dataset, both in spanish (spa)
and in english (eng).

These datasets are just examples, and shall be used only to
test the application in a very high level-small scope situation,
and as a "template", providing a working example of how the
text lines shall be.