Print recipe title capitalized
Recipe title shall be printed on screen and also on the
notes file with the first letter in Uppercase (capitalized).

This is a rather "stylish" change, but its rationale is also
technical. The "recipes.csv" is all in lowercase (except for
the URL field that may contain uppercase letters), so when
user wants to search for recipes with specific titles, it does
not matter if the user inputs "eggs", "Eggs", or "eGGs", because
the input is going to be set to all lowercase letters. This ensures
that all recipes with "eggs" will be printed.

However, when printing the recipe title, if first letter is not
in uppercase it is kind of "ugly". So this commit fixes that,
without affecting the remain functionality.